+++ src/oms_sensemaking/geospatial/controllers.py
@@ -22,6 +22,7 @@ from oms_sensemaking.core.controllers import SensemakerController
 from oms_sensemaking.core.error_loggers import BaseErrorLogger
 from oms_sensemaking.core.events import AuditLogEvent, AuditLogEventConsumer, EventFilter
 from oms_sensemaking.core.exceptions import TrackLengthError
+from oms_sensemaking.core.observability import with_metrics_collection
 from oms_sensemaking.dao.track import APITrack
 from oms_sensemaking.geospatial.schemas import GeospatialSensemakerConfig
 from oms_sensemaking.geospatial.sensemakers import CotravelSensemaker, LoiterSensemaker, SimilarTracksSensemaker
@@ -107,6 +108,7 @@ class GeospatialSensemakerController(SensemakerController):
 
         super().stop()
 
+    @with_metrics_collection
     def _ensure_uuid(self, id) -> UUID:
         """
         Function used to ensure that ids used
@@ -117,6 +119,7 @@ class GeospatialSensemakerController(SensemakerController):
         else:
             return UUID(id)
 
+    @with_metrics_collection
     def handle_event(self, event: AuditLogEvent) -> bool:
         """
         Handle inbound OMS event.