+++ CHANGELOG.md
@@ -59,6 +59,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Log request header information
 - Added authorization for POST /aac/clear endpoint
+- OpenTelemetry integration for queue processing time and event counting metrics
 
 ### Changed
 - Dockerfile and related build files are now ready to build UBI8 based images for Sensemaking in the AIDE environment